documents-api/pyproject.toml
[tool.poetry]
name = "documents-api"
version = "0.1.0"
description = "Documents API service for document processing and analysis"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
pypdf2 = "^3.0.1"
requests = "^2.32.3"
asyncpg = "^0.29.0"
google-generativeai = "^0.8.5"
pydantic = "^2.7.4"
starlette = "^0.37.2"
pillow = "^11.2.1"
pyyaml = "^6.0.2"
celery = "^5.3.4"
redis = "^5.0.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

